pyproject.toml
[tool.poetry]
name = "jaxfluids"
version = "0.1.0"
description = "Fully-differentiable CFD solver for compressible two-phase flows."
authors = [
    "Deniz Bezgin <deniz.bezgin@tum.de>",
    "Aaron Buhendwa <aaron.buhendwa@tum.de>"
]
license = "GNU GPLv3"
readme = "README.md"
packages = [{include = "jaxfluids"}]

[tool.poetry.dependencies]
python = "^3.8"
jax = "^0.4.6"
dm-haiku = "^0.0.7"
h5py = "^3.8.0"
numpy = "^1.24.2"
optax = "^0.1.4"

[tool.poetry.group.dev.dependencies]
black = "23.3.0"
pre-commit = "^3.2.1"
ruff = "^0.0.260"
isort = "^5.10.0"

[tool.ruff]
line-length = 100
ignore = ["D203"]

[tool.black]
line-length = 100
target-version = ['py310']
force-exclude = '''
    docs/
    examples/
    notebooks/
    tutorials/
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
